move integration tests to convert_itest
